plan, engine: get rid of trim sorts optimizer
