+++ logicalplan/plan_test.go
@@ -138,6 +138,50 @@ func TestMatcherPropagation(t *testing.T) {
 	}
 }
 
+func TestTrimSorts(t *testing.T) {
+	cases := []struct {
+		name     string
+		expr     string
+		expected string
+	}{
+		// this test case is ok since the engine determines sorting order
+		// before running optimziers
+		{
+			name:     "simple sort",
+			expr:     "sort(X)",
+			expected: "X",
+		},
+		{
+			name:     "sort",
+			expr:     "sum(sort(X))",
+			expected: "sum(X)",
+		},
+		{
+			name:     "nested",
+			expr:     "sum(sort(rate(X[1m])))",
+			expected: "sum(rate(X[1m]))",
+		},
+		{
+			name:     "weirdly nested",
+			expr:     "sum(sort(sqrt(sort(X))))",
+			expected: "sum(sqrt(X))",
+		},
+		{
+			name:     "sort in binary expression",
+			expr:     "sort(sort(sqrt(X))/sort(sqrt(Y)))",
+			expected: "sqrt(X) / sqrt(Y)",
+		},
+	}
+	for _, tcase := range cases {
+		t.Run(tcase.name, func(t *testing.T) {
+			expr, err := parser.ParseExpr(tcase.expr)
+			testutil.Ok(t, err)
+
+			testutil.Equals(t, tcase.expected, trimSorts(expr).String())
+		})
+	}
+}
+
 func cleanUp(replacements map[string]*regexp.Regexp, expr string) string {
 	for replacement, match := range replacements {
 		expr = match.ReplaceAllString(expr, replacement)
